migraine_backend/app/utils/jwt.py
from datetime import datetime, timedelta
from jose import JWTError, jwt

SECRET_KEY = "migraine_super_secret_key"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60

def create_access_token(data: dict, expires_delta: timedelta = None):
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

def decode_access_token(token: str):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None
def verify_access_token(token: str):
    payload = decode_access_token(token)
    if payload is None or "user_id" not in payload:
        raise JWTError("Geçersiz kimlik doğrulama tokeni.")
    return payload
