requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.2.1
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
# ML Libraries
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
imbalanced-learn==0.11.0